requirements.txt
google-genai==1.29.0
wordfreq==3.1.1
nltk==3.9.1
